requirements.txt
opencv-python==4.10.0.84
numpy==2.2.1
fastapi==0.115.6
python-jose==3.3.0
pydantic==2.10.4
ultralytics==8.3.55
uvicorn==0.34.0
onnxruntime==1.20.1